dailytask/requirements.txt
tensorflow
streamlit
numpy
pillow
gdown